refactor(mask): move all mask code into the inputMaskFilter file - since we extend, we can just override necessary pieces
