+++ aurelia-slickgrid/src/aurelia-slickgrid/filters/inputFilter.ts
@@ -12,10 +12,9 @@ import {
 import * as $ from 'jquery';
 
 export class InputFilter implements Filter {
-  private _clearFilterTriggered = false;
-  private _inputType = 'text';
-  private _hasMask = false;
-  private $filterElm: any;
+  protected _clearFilterTriggered = false;
+  protected _inputType = 'text';
+  protected $filterElm: any;
   grid: any;
   searchTerms: SearchTerm[];
   columnDef: Column;
@@ -24,16 +23,6 @@ export class InputFilter implements Filter {
   /** Getter for the Filter Operator */
   get columnFilter(): ColumnFilter {
     return this.columnDef && this.columnDef.filter || {};
-  }
-
-  /** Getter of the input mask, when provided */
-  get inputMask(): string {
-    return this.columnDef.params && this.columnDef.params && this.columnDef.params.mask;
-  }
-
-  /** Setter for knowing if input has a mask filter provided */
-  set hasMask(hasMask: boolean) {
-    this._hasMask = hasMask;
   }
 
   /** Getter of input type (text, number, password) */
@@ -52,7 +41,7 @@ export class InputFilter implements Filter {
   }
 
   /** Getter for the Grid Options pulled through the Grid Object */
-  private get gridOptions(): GridOption {
+  protected get gridOptions(): GridOption {
     return (this.grid && this.grid.getOptions) ? this.grid.getOptions() : {};
   }
 
@@ -80,23 +69,8 @@ export class InputFilter implements Filter {
     // step 3, subscribe to the keyup event and run the callback when that happens
     // also add/remove "filled" class for styling purposes
     this.$filterElm.on('keyup input change', (e: any) => {
-      let value = '';
-      if (e && e.target && e.target.value) {
-        value = e.target.value;
+      const value = e && e.target && e.target.value || '';
 
-        // if it has a mask, we need to do a bit more work
-        // and replace the filter string by the masked output without triggering an event
-        if (this._hasMask) {
-          const unmaskedValue = this.unmaskValue(value);
-          const maskedValue = this.maskValue(unmaskedValue);
-          value = unmaskedValue;
-
-          if (e.keyCode >= 48) {
-            this.$filterElm.val(maskedValue); // replace filter string with masked string
-            e.preventDefault();
-          }
-        }
-      }
       if (this._clearFilterTriggered) {
         this.callback(e, { columnDef: this.columnDef, clearFilterTriggered: this._clearFilterTriggered });
         this._clearFilterTriggered = false; // reset flag for next use
@@ -139,13 +113,13 @@ export class InputFilter implements Filter {
   }
 
   //
-  // private functions
+  // protected functions
   // ------------------
 
   /**
    * Create the HTML template as a string
    */
-  private buildTemplateHtmlString() {
+  protected buildTemplateHtmlString() {
     const columnId = this.columnDef && this.columnDef.id;
     let placeholder = (this.gridOptions) ? (this.gridOptions.defaultFilterPlaceholder || '') : '';
     if (this.columnFilter && this.columnFilter.placeholder) {
@@ -158,7 +132,7 @@ export class InputFilter implements Filter {
    * From the html template string, create a DOM element
    * @param filterTemplate
    */
-  private createDomElement(filterTemplate: string, searchTerm?: SearchTerm) {
+  protected createDomElement(filterTemplate: string, searchTerm?: SearchTerm) {
     const columnId = this.columnDef && this.columnDef.id;
     const $headerElm = this.grid.getHeaderRowColumn(columnId);
     $($headerElm).empty();
@@ -182,45 +156,5 @@ export class InputFilter implements Filter {
     }
 
     return $filterElm;
-  }
-
-  /** From a regular string, we will use the mask to output a new string */
-  private maskValue(inputValue: string): string {
-    let i = 0;
-    let maskedValue = '';
-
-    if (this.inputMask) {
-      maskedValue = this.inputMask.replace(/[09A]/gi, (match) => {
-        // only replace the char when the mask is a 0 or 9 for a digit OR the mask is "A" and the char is a non-digit meaning a string char
-        if (
-          ((match === '0' || match === '9') && /\d*/g.test(inputValue[i]))    // mask is 0 or 9 and value is a digit
-          || (match.toUpperCase() === 'A' && /[^\d]*/gi.test(inputValue[i]))  // OR mask is an "A" and value is non-digit
-        ) {
-          return inputValue[i++] || '';
-        }
-        return '';
-      });
-    }
-
-    return maskedValue;
-  }
-
-  /** From a masked string, we will remove the mask and make a regular string again */
-  private unmaskValue(maskedValue: string): string {
-    let maskedMatches = [];
-
-    if (/[a-z]/gi.test(this.inputMask) && /[^0-9]/gi.test(this.inputMask)) {
-      // chars only without numbers in the string
-      maskedMatches = maskedValue.match(/[a-z]*/gi);
-    } else if (/[0-9]/gi.test(this.inputMask) && /[^a-z]/gi.test(this.inputMask)) {
-      // numbers only without chars in the string
-      maskedMatches = maskedValue.match(/[0-9]*/gi);
-    } else {
-      maskedMatches = maskedValue.match(/[0-9a-z]*/gi);
-    }
-    if (Array.isArray(maskedMatches)) {
-      return maskedMatches.join('');
-    }
-    return '';
   }
 }
